fix: parsing `sort` input schema
There was an issue, because prisma cannot handle multiple order values. So from now on, I reduce the original sort input schema and take only the first entry of it - which is a bit hacky, but I consider it as the best solution for now since I only have use cases where only one order value matters.

- added `parseSortToOrderBy` util
- updated `toggleSortParam` inside the `useFilterParams` custom hook to handle only a single sort value
- updated previous sort handlers
- added `parseSortToOrderBy` util where needed