chore: prevent throwing context error
There are cases when I want to use the `useSidebar` hook outside of a `SidebarProvider`. I added an extra `provideFallback` param to the hook to prevent throwing an error when it's being called outside of its context provider. In this case, a `CONTEXT_FALLBACK` constant will be provided instead of giving back the actual sidebar context.