fix: uploading collection
Form data couldn't be sent to the `createCollection` server action using `form.getValues()`. I refactored this part of the code a bit, and now it's working correctly.

Note!: awaiting `startUpload`, the result can be `undefined`. I cannot find anything related on the docs, so I expect that if the result is `undefined` === no files have been uploaded -> which means, deleting files is not necessary if request fails